requirements.txt
cloudinary
Flask-Login
Flask-SQLAlchemy
Flask-WTF
bootstrap-flask
Flask-Migrate
pymysql[rsa]
randimage
SQLAlchemy
Werkzeug
gunicorn
python-dotenv
pandas